Epub reader based on Readium for picture books
A simple epub reader based on Readium project, only designed for
picture book (fixed layout), code modified to communicate with a CMS so
that reader can download epub file from remote server.
Project developed according to the request of a public library for
non-profit educational purpose, this is not a commercial bookstore.